+++ apps/backoffice/src/layouts/page-layout.tsx
@@ -12,6 +12,8 @@ export const PageLayout = ({
 }: PageLayoutProps) => {
   return (
     <Box
+      display="flex"
+      flexDirection="column"
       padding={3}
       width={isFullWidth ? "100%" : "60%"}
       marginX={isFullWidth ? "" : "auto"}