+++ apps/backoffice/src/pages/index.tsx
@@ -1,14 +1,34 @@
+import { AccessControl } from "@/components/access-control";
+import { CardDetails } from "@/components/cards";
 import { PageHeader } from "@/components/headers";
+import { SubscriptionKeys } from "@/generated/api/SubscriptionKeys";
+import useFetch from "@/hooks/use-fetch";
 import { AppLayout, PageLayout } from "@/layouts";
+import { hasRequiredAuthorizations } from "@/utils/auth-util";
+import { Grid } from "@mui/material";
+import { useSession } from "next-auth/react";
 import { useTranslation } from "next-i18next";
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
-import { ReactElement } from "react";
+import { ReactElement, useEffect } from "react";
 
 const pageTitleLocaleKey = "routes.overview.title";
 const pageDescriptionLocaleKey = "routes.overview.description";
 
 export default function Home() {
   const { t } = useTranslation();
+  const { data: session } = useSession();
+  const { data: mkData, fetchData: mkFetchData } = useFetch<SubscriptionKeys>();
+
+  useEffect(() => {
+    if (
+      hasRequiredAuthorizations(session, {
+        requiredPermissions: ["apiservicewrite"]
+      })
+    ) {
+      mkFetchData("getManageKeys", {}, SubscriptionKeys);
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
 
   return (
     <>
@@ -16,6 +36,28 @@ export default function Home() {
         title={pageTitleLocaleKey}
         description={pageDescriptionLocaleKey}
       />
+      <Grid container spacing={2}>
+        <Grid item xs={12} sm={8} md={8} lg={6} xl={6}>
+          <AccessControl requiredPermissions={["apiservicewrite"]}>
+            <CardDetails
+              title="routes.keys.manage.title"
+              cta={{ label: "routes.keys.manage.shortcut", href: "/keys" }}
+              rows={[
+                {
+                  label: "keys.primary.title",
+                  value: mkData?.primary_key,
+                  kind: "apikey"
+                },
+                {
+                  label: "keys.secondary.title",
+                  value: mkData?.secondary_key,
+                  kind: "apikey"
+                }
+              ]}
+            />
+          </AccessControl>
+        </Grid>
+      </Grid>
     </>
   );
 }